requirements.txt
rich
langchain
huggingface_hub
loguru
pydantic
python-dotenv
openai
tiktoken
inquirer
prompt_toolkit